requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
sqlalchemy==2.0.35
aiosqlite==0.20.0
anthropic==0.34.2
google-generativeai==0.8.3
python-dotenv==1.0.1
pillow==10.4.0
aiofiles==23.2.1
